+++ src/CLIError.js
@@ -1,4 +1,0 @@
-// @flow
-
-export default class CLIError extends Error {}
-